Moved KinematicElement to its own header
